move SpacedTokens to utils
