+++ app/src/main/java/helium314/keyboard/latin/utils/SpacedTokens.kt
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
-package helium314.keyboard.latin
+package helium314.keyboard.latin.utils
 
 /**
  * Tokenizes strings by groupings of non-space characters, making them iterable. Note that letters,
